New class for logging
The Logging class should be used for logging useful information (error, warnings, informative messages and debug strings). It supports arbitrary nesting levels of indentation, multiple message types, and the log messages can be sent to any output stream (a string, a file, or the terminal -- the latter is the default).