SQLite3Connection objects do not create empty databases.
The behaviour of SQLite3Connection was to silently create a new database
if the file specified in the constructor's parameter was not found. This
is no longer the case: if the file does not exist, an exception is
thrown. (This assumes that one never wants to modify a SQLite3 database,
which is the case of this module: we're only using SQLite3Connection to
query UCDS databases.)