Dipole parameters are no longer constant, "debug" mode for dipoleFit.
A new structure, Dipole_parameters_t, has been added to the codebase. It
is used by PlanckVelocity and daCapo, and it reads the parameters from a
configuration file. The way dipole_fit works has been changed
accordingly.

run_dipole_fit has been restructured in order to be able to write TODs
containing the convolved dipole and the galactic sidelobes into FITS
files (this feature is mainly useful for debugging).